*** empty log message ***

+++ web_module/branches/intro_event_2007_03_28/src/main/com/topcoder/web/common/dao/ImageDAO.java
@@ -12,5 +12,5 @@ public interface ImageDAO {
 
     List<Image> getImages(Integer typeId);
 
-    Image find(Integer id);
+    Image find(Long id);
 }
