+++ web_module/branches/intro_event_2007_03_28/src/main/com/topcoder/web/common/dao/hibernate/ImageDAOHibernate.java
@@ -25,7 +25,7 @@ public class ImageDAOHibernate extends Base implements ImageDAO {
         return findAll(TimeZone.class);
     }
 
-    public Image find(Integer id) {
+    public Image find(Long id) {
         return (Image) find(Image.class, id);
     }
